+++ packages/@expo/metro-config/src/__tests__/babel-transformer.test.ts
@@ -78,6 +78,7 @@ it(`passes the environment as isServer to the babel preset`, () => {
         bundler: 'metro',
         name: 'metro',
         platform: 'ios',
+        projectRoot: expect.any(String),
       },
       cloneInputAst: false,
       code: false,